+++ lib/vagrant/box.rb
@@ -57,6 +57,7 @@ module Vagrant
     # method requires that `name` and `uri` be set. The logic of this method
     # is kicked out to the `box_add` registered middleware.
     def add
+      raise Errors::BoxAlreadyExists.new(:name => name) if File.directory?(directory)
       env.actions.run(:box_add, { "box" => self })
     end