+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.6.4 (unreleased)
 
+  - Box add checks if a box already exists before the download. [GH-170]
   - NFS no longer attempts to clean exports file if VM is not created,
     which was causing a stack trace during recovery. [related to GH-166]
   - Basic validation added for Chef configuration (both solo and server).